bhxh-tool-enhanced-python/requirements.txt
# Enhanced BHXH Tool - Python Dependencies
# Version: 2.0.0

# HTTP requests and form handling
requests>=2.31.0
httpx>=0.25.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Excel processing
openpyxl>=3.1.0
pandas>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Caching
cachetools>=5.3.0
diskcache>=5.6.0

# Logging (built-in, but enhanced)
structlog>=23.0.0

# Async support
asyncio-throttle>=1.0.0

# Date/time handling (built-in datetime, but enhanced)
python-dateutil>=2.8.0

# Utilities
click>=8.1.0  # For CLI interface
rich>=13.0.0  # For enhanced console output

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0